Fix compatibility issues with Rails 5.1
In Rails 5.1, #extract_schema_qualified_name is redefined in the
PostgreSQLAdapter.  Because CockroachDBAdapter defines it slightly
differently, this causes compatibility issues between the two.  This
change ensures that the namespaces are no longer colliding.

This also restores the #index_name_length and #table_alias_length methods which
were changed in ActiveRecord 5.1 to use `SHOW max_identifier_length` which does not
exist in CockroachDB.  We hardcode this to 63 (as it was previously in
ActiveRecord 5.0).