schema.rb generation: ensure tables are dumped when indexes are used
Ensure `db/schema.rb` is generated properly when the database uses
indexes.

Previous to this patch the code was written to work with ActiveRecord
v5 which has a different ActiveRecord::ConnectionAdapters::IndexDefinition
struct compared to ActiveRecord v4.

The usage of the v5 struct lead the `db/schema.rb` to skip tables with
indexes and leave just this comment:

```
 # Could not dump table "<TABLE>" because of following ArgumentError
 #   struct size differs
```

Signed-off-by: Flavio Castelli <fcastelli@suse.com>