refactor(Gemfile): Simplify rails loading
The `RAILS_SOURCE` env was not documented, and the methods were
not up-to-date (and crashing). I think this is not necessary as
we can manually change the `gem` specification for testing. I
added a related comment in the Gemfile.

`RAILS_VERSION` has basically the same problem: we can manually
change the tag if we need to make a specific test.

As the code to fetch the latest version of Rails wasn't working
anymore, I changed it to something more robust, taking advantage
of the `Gem::Version` and `Gem::Requirement` classes. And without
relying on `eval`