feat: Support Rails 7.1
This PR includes loads of small changes that won't be
mentionned here to adapt to the new Rails code.

Loads of tests ignored tests have been added back as
they were not error prone anymore. There are likely
more of these.

We removed the deprecated `configure_connection` method
override. It is now unnecessary as CockroachDB supports
`SET TIMEZONE <...>`.

We dropped support for CockroachDB version before 22.X.X.

We removed a lot of copy-pasted code in the test thanks to
new file separation in Rails 7.1 test configuration.

Signed-off-by: Ulysse Buonomo <buonomo.ulysse@gmail.com>