fix: Refactor new connection
The old method was calling `AbstractAdapter#initialize`
with a connection object, rather than a hash. This was
then setting `@unconfigured_connection`, which in turns
forced to re-run `#configure_connection` when a first
query was made. This is usually OK (although unnecessary)
except if `use_follower_reads_for_type_introspection` is
used, since it sets AOST for a single query, which fails
within a transaction.

Fixes #320