fix: do not alter user comments
We used to have a regex that removed the first and last
single quotes from a comment. There is no information
about this decision, nor any related test. This might
have been a difference between PostgreSQL and CockroachDB.

Since tests are passing without it, and comments are not
quoted, we can remove the regex. Moreover, if a comment would
start or end with a single quote, this would remove that
quote, which is not desirable.

Fixes #381