fix(tests): clear query cache on insert
Since [rails pull request #52428][1], `#execute_batch` does not trigger
a cache clear anymore. However, `#insert_fixtures_set` relies on that
clear to ensure consistency. In the postgresql adapter, this is ensured
by a call to `#execute` rather than `#execute_batch` in
`#disable_referential_integrity`. Since we are not always calling
`#disable_referential_integrity`, we need to ensure that the cache is
cleared when running our statements by calling `#execute` instead of
`#execute_batch`.

[1]: https://github.com/rails/rails/pull/52428

Signed-off-by: Ulysse Buonomo <buonomo.ulysse@gmail.com>